requirements.txt
pytest
pytest_voluptuous
requests
python-dotenv
voluptuous
Faker
selene==2.0.0b6
allure-pytest